+++ src/mainwindow.cpp
@@ -635,6 +635,10 @@ void MainWindow::setCurrentNote( Note note, bool updateNoteText, bool updateSele
     this->currentNote = note;
     QString name = note.getName();
 
+    // set the note text edit to readonly if note file is not writeable
+    QFileInfo *f = new QFileInfo( this->notesPath + QDir::separator() + note.getFileName() );
+    ui->noteTextEdit->setReadOnly( !f->isWritable() );
+
     // find and set the current item
     if ( updateSelectedNote )
     {