+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # QOwnNotes Changelog
 
 ## 0.43
-- added functionality restore a trashed note directly on the ownCloud server, restoring the note's versions with it. (needs QOwnNotesAPI v0.4)
-- updating to QOwnNotesAPI v0.4 is suggested, because it also contains two imporant bugfixes for the trashed notes api
+- added functionality to restore a trashed note directly on the ownCloud server, restoring the note's versions with it. (needs QOwnNotesAPI v0.4)
+- updating to QOwnNotesAPI v0.4 is suggested, because it also contains two important bugfixes for the trashed notes api
 - fixed a bug when the ownCloud path and the notes path are the same
 - added write permissions to the demo notes
 - note editing is now disallowed if note file is readonly